sweep pkg/util for easy moves
