+++ pkg/oc/lib/describe/chaindescriber.go
@@ -15,12 +15,12 @@ import (
 
 	imagev1 "github.com/openshift/api/image/v1"
 	buildv1client "github.com/openshift/client-go/build/clientset/versioned/typed/build/v1"
+	dotutil "github.com/openshift/origin/pkg/oc/lib/dot"
 	buildedges "github.com/openshift/origin/pkg/oc/lib/graph/buildgraph"
 	buildanalysis "github.com/openshift/origin/pkg/oc/lib/graph/buildgraph/analysis"
 	buildgraph "github.com/openshift/origin/pkg/oc/lib/graph/buildgraph/nodes"
 	osgraph "github.com/openshift/origin/pkg/oc/lib/graph/genericgraph"
 	imagegraph "github.com/openshift/origin/pkg/oc/lib/graph/imagegraph/nodes"
-	dotutil "github.com/openshift/origin/pkg/util/dot"
 	"github.com/openshift/origin/pkg/util/parallel"
 )
