+++ pkg/oc/lib/dot/dot_test.go
