+++ pkg/oc/lib/newapp/portutils/common.go
