--- pkg/util/httpproxy/doc.go
@@ -1,2 +1,0 @@
-// Package httpproxy contains an upgrade-aware HTTP single-host reverse proxy.
-package httpproxy
