--- pkg/util/rest/mapper.go
@@ -1,1 +1,0 @@
-package rest
